[nexus] Clean up API versioning: chain conversions, fix module docs
This PR refactors a set of versioned type conversions to properly follow the
[API versioning guide's](https://github.com/oxidecomputer/dropshot-api-manager/blob/main/guides/new-version.md)
recommendations, including chaining through consecutive versions, ordering
conventions, and aligning types to correctly match the ranges and
where they should be defined.

Changes:
  - Chain InstanceCreate conversions: v2025122300 → v2026010300 →
    v2026010500 → params
  - Update endpoint handlers to delegate through the version chain
  - Add per-section validity docs for modules where types have different
    supersession versions
  - Remove empty v2026011501.rs after types moved to v2026010500
