+++ nexus/external-api/src/v2025112000.rs
@@ -2,9 +2,10 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//! Nexus external types that changed from 2025112000 to 2025120300
+//! Types from API version 2025112000 (INITIAL) that changed in version
+//! 2025120300 (LOCAL_STORAGE).
 
-use crate::v2025121200;
+use crate::v2025120300;
 use crate::v2026010100;
 use nexus_types::external_api::params;
 use omicron_common::api::external;
@@ -220,9 +221,9 @@ pub struct InstanceCreate {
     /// By default, all instances have outbound connectivity, but no inbound
     /// connectivity. These external addresses can be used to provide a fixed,
     /// known IP address for making inbound connections to the instance.
-    // Delegates through v2025121200 → params::ExternalIpCreate
+    // Delegates through v2025120300 → params::ExternalIpCreate
     #[serde(default)]
-    pub external_ips: Vec<v2025121200::ExternalIpCreate>,
+    pub external_ips: Vec<v2025120300::ExternalIpCreate>,
 
     /// The multicast groups this instance should join.
     ///
@@ -302,9 +303,9 @@ pub struct InstanceCreate {
     pub cpu_platform: Option<external::InstanceCpuPlatform>,
 }
 
-impl From<InstanceCreate> for v2025121200::InstanceCreate {
-    fn from(old: InstanceCreate) -> v2025121200::InstanceCreate {
-        v2025121200::InstanceCreate {
+impl From<InstanceCreate> for v2025120300::InstanceCreate {
+    fn from(old: InstanceCreate) -> v2025120300::InstanceCreate {
+        v2025120300::InstanceCreate {
             identity: old.identity,
             ncpus: old.ncpus,
             memory: old.memory,
