+++ nexus/external-api/src/v2025121200.rs
@@ -2,35 +2,27 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//! Nexus external types that changed from 2025121200 to 2026010100.
+//! Types from API version 2025121200 (BGP_PEER_COLLISION_STATE) that changed in
+//! subsequent versions.
 //!
-//! Version 2025121200 types (before `ip_version` preference was added for
-//! default IP pool selection).
+//! Types before `ip_version` preference was added for default IP pool selection.
 //!
-//! ## IP Pool Selection Changes
+//! ## IP Pool Selection Types
 //!
-//! Key differences from newer API versions:
-//! - [`FloatingIpCreate`], [`EphemeralIpCreate`], and [`ExternalIpCreate`]
-//!   don't have the `ip_version` field. Newer versions allow specifying
-//!   IPv4/IPv6 preference when allocating from default pools.
-//! - When multiple default pools of different IP versions exist for a silo,
-//!   older clients cannot resolve the conflict. Newer API versions
-//!   require the `ip_version` field in this scenario.
+//! Valid until 2025122300 (IP_VERSION_AND_MULTIPLE_DEFAULT_POOLS).
 //!
-//! Affected endpoints:
-//! - `POST /v1/floating-ips` (floating_ip_create)
-//! - `POST /v1/instances/{instance}/external-ips/ephemeral` (instance_ephemeral_ip_attach)
-//! - `POST /v1/instances` (instance_create)
+//! [`FloatingIpCreate`] and [`EphemeralIpCreate`] don't have the `ip_version`
+//! field. Newer versions allow specifying IPv4/IPv6 preference when allocating
+//! from default pools.
 //!
 //! ## Multicast Types
 //!
+//! Valid until 2026010800 (MULTICAST_IMPLICIT_LIFECYCLE_UPDATES).
+//!
 //! Multicast types are re-exported from `v2025122300`.
-//! Both versions use `NameOrId` for group references and have the same
-//! explicit create/update semantics.
 //!
 //! [`FloatingIpCreate`]: self::FloatingIpCreate
 //! [`EphemeralIpCreate`]: self::EphemeralIpCreate
-//! [`ExternalIpCreate`]: self::ExternalIpCreate
 
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -38,7 +30,7 @@ use serde::{Deserialize, Serialize};
 use nexus_types::external_api::params;
 use omicron_common::api::external;
 
-use crate::{v2026010100, v2026010300};
+use crate::v2026010300;
 
 // Re-export multicast types from v2025122300.
 // They're identical for both versions (both use NameOrId, explicit
@@ -69,42 +61,6 @@ impl From<EphemeralIpCreate> for params::EphemeralIpCreate {
     }
 }
 
-/// The type of IP address to attach to an instance during creation.
-#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum ExternalIpCreate {
-    /// An IP address providing both inbound and outbound access.
-    /// The address is automatically assigned from the provided IP pool
-    /// or the default IP pool if not specified.
-    Ephemeral {
-        /// Name or ID of the IP pool to use. If unspecified, the
-        /// default IP pool will be used.
-        pool: Option<external::NameOrId>,
-    },
-    /// A floating IP address.
-    Floating {
-        /// The name or ID of the floating IP address to attach.
-        floating_ip: external::NameOrId,
-    },
-}
-
-// Converts to v2026010300::ExternalIpCreate (adds ip_version: None)
-impl From<ExternalIpCreate> for v2026010300::ExternalIpCreate {
-    fn from(old: ExternalIpCreate) -> v2026010300::ExternalIpCreate {
-        match old {
-            ExternalIpCreate::Ephemeral { pool } => {
-                v2026010300::ExternalIpCreate::Ephemeral {
-                    pool,
-                    ip_version: None,
-                }
-            }
-            ExternalIpCreate::Floating { floating_ip } => {
-                v2026010300::ExternalIpCreate::Floating { floating_ip }
-            }
-        }
-    }
-}
-
 /// Parameters for creating a new floating IP address for instances.
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct FloatingIpCreate {
@@ -128,77 +84,6 @@ impl From<FloatingIpCreate> for v2026010300::FloatingIpCreate {
             ip: old.ip,
             pool: old.pool,
             ip_version: None,
-        }
-    }
-}
-
-/// Create-time parameters for an `Instance`
-#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
-pub struct InstanceCreate {
-    #[serde(flatten)]
-    pub identity: external::IdentityMetadataCreateParams,
-    /// The number of vCPUs to be allocated to the instance
-    pub ncpus: external::InstanceCpuCount,
-    /// The amount of RAM (in bytes) to be allocated to the instance
-    pub memory: external::ByteCount,
-    /// The hostname to be assigned to the instance
-    pub hostname: external::Hostname,
-    /// User data for instance initialization systems (such as cloud-init).
-    #[serde(default, with = "params::UserData")]
-    pub user_data: Vec<u8>,
-    /// The network interfaces to be created for this instance.
-    #[serde(default)]
-    pub network_interfaces: v2026010100::InstanceNetworkInterfaceAttachment,
-    /// The external IP addresses provided to this instance.
-    #[serde(default)]
-    pub external_ips: Vec<ExternalIpCreate>,
-    /// The multicast groups this instance should join.
-    #[serde(default)]
-    pub multicast_groups: Vec<external::NameOrId>,
-    /// A list of disks to be attached to the instance.
-    #[serde(default)]
-    pub disks: Vec<params::InstanceDiskAttachment>,
-    /// The disk the instance is configured to boot from.
-    #[serde(default)]
-    pub boot_disk: Option<params::InstanceDiskAttachment>,
-    /// An allowlist of SSH public keys to be transferred to the instance.
-    pub ssh_public_keys: Option<Vec<external::NameOrId>>,
-    /// Should this instance be started upon creation; true by default.
-    #[serde(default = "params::bool_true")]
-    pub start: bool,
-    /// The auto-restart policy for this instance.
-    #[serde(default)]
-    pub auto_restart_policy: Option<external::InstanceAutoRestartPolicy>,
-    /// Anti-Affinity groups which this instance should be added.
-    #[serde(default)]
-    pub anti_affinity_groups: Vec<external::NameOrId>,
-    /// The CPU platform to be used for this instance.
-    #[serde(default)]
-    pub cpu_platform: Option<external::InstanceCpuPlatform>,
-}
-
-impl From<InstanceCreate> for v2026010100::InstanceCreate {
-    fn from(old: InstanceCreate) -> v2026010100::InstanceCreate {
-        v2026010100::InstanceCreate {
-            identity: old.identity,
-            ncpus: old.ncpus,
-            memory: old.memory,
-            hostname: old.hostname,
-            user_data: old.user_data,
-            network_interfaces: old.network_interfaces,
-            external_ips: old
-                .external_ips
-                .into_iter()
-                .map(Into::into)
-                .collect(),
-            multicast_groups: old.multicast_groups,
-            disks: old.disks,
-            boot_disk: old.boot_disk,
-            ssh_public_keys: old.ssh_public_keys,
-            start: old.start,
-            auto_restart_policy: old.auto_restart_policy,
-            anti_affinity_groups: old.anti_affinity_groups,
-            cpu_platform: old.cpu_platform,
         }
     }
 }
